More work on #16
This commit updates all GUICtrlCreateLabel() functions to save the ID into a global variable, so that the labels can be enabled/disabled along with other GUI elements. This commit also saves all GUICrtlCreateTabItem() ID into a global variable, so that later functions can act on the active tab, I think.